requirements.txt
python-telegram-bot[job-queue]==21.0.1
google-generativeai==0.3.2
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
Pillow==10.0.1
reportlab==4.0.4
asyncio-mqtt==0.13.0
aiofiles==23.2.1
requests==2.31.0